Remove webpack-blocks
