+++ package.json
@@ -24,8 +24,7 @@
     "babel-core": "~6"
   },
   "devDependencies": {
-    "@webpack-blocks/webpack2": "^0.4.0",
-    "babel-loader": "^6.4.0",
+    "babel-loader": "^7.1.4",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-preset-es2015": "^6.22.0",
     "eslint": "^3.17.1",
@@ -35,10 +34,11 @@
     "jest": "^22.4.3",
     "mocha": "^3.2.0",
     "source-map-support": "^0.4.14",
-    "webpack": "^3.0.0"
+    "webpack": "^4.5.0",
+    "webpack-cli": "^2.0.14"
   },
   "peerDependencies": {
-    "webpack": "^1 || ^2 || ^3"
+    "webpack": "^1 || ^2 || ^3 || ^4"
   },
   "keywords": [
     "webpack",
